+++ src/apps/chifra/cmd/help_text.go
@@ -31,7 +31,7 @@ const helpText = `  Accounts:
     init          initialize the TrueBlocks system by downloading from IPFS
   Other:
     explore       open a local or remote explorer for one or more addresses, blocks, or transactions
-    slurp         fetch data from EtherScan for any address
+    slurp         fetch data from Etherscan for any address
   Flags:
     -h, --help    display this help screen