+++ src/apps/chifra/internal/export/validate.go
@@ -117,7 +117,7 @@ func (opts *ExportOptions) validateExport() error {
 	}
 
 	if !validate.CanArticulate(opts.Articulate) {
-		return validate.Usage("The {0} option requires an EtherScan API key.", "--articulate")
+		return validate.Usage("The {0} option requires an Etherscan API key.", "--articulate")
 	}
 
 	// Note that this does not return if the index is not initialized