+++ src/apps/chifra/internal/slurp/validate.go
@@ -5,6 +5,9 @@
 package slurpPkg
 
 import (
+	"errors"
+	"fmt"
+
 	"github.com/TrueBlocks/trueblocks-core/src/apps/chifra/pkg/validate"
 )
 
@@ -27,6 +30,33 @@ func (opts *SlurpOptions) validateSlurp() error {
 
 	if opts.Globals.Chain != "mainnet" {
 		return validate.Usage("The {0} command is currently available only on the {1} chain.", "slurp", "mainnet")
+	}
+
+	if opts.Sleep < .25 {
+		return validate.Usage("The {0} option ({1}) must {2}.", "--sleep", fmt.Sprintf("%f", opts.Sleep), "be at least .25")
+	}
+
+	if opts.PerPage < 10 {
+		return validate.Usage("The {0} option ({1}) must {2}.", "--per_page", fmt.Sprintf("%d", opts.PerPage), "be at least 10")
+	} else if opts.PerPage > 10000 {
+		return validate.Usage("The {0} option ({1}) must {2}.", "--per_page", fmt.Sprintf("%d", opts.PerPage), "be no more than 10,000")
+	}
+
+	err = validate.ValidateIdentifiers(
+		opts.Globals.Chain,
+		opts.Blocks,
+		validate.ValidBlockIdWithRange,
+		1,
+		&opts.BlockIds,
+	)
+	if err != nil {
+		if invalidLiteral, ok := err.(*validate.InvalidIdentifierLiteralError); ok {
+			return invalidLiteral
+		}
+		if errors.Is(err, validate.ErrTooManyRanges) {
+			return validate.Usage("Specify only a single block range at a time.")
+		}
+		return err
 	}
 
 	return opts.Globals.Validate()