+++ src/apps/chifra/internal/traces/validate.go
@@ -30,7 +30,7 @@ func (opts *TracesOptions) validateTraces() error {
 		}
 
 		if !validate.CanArticulate(opts.Articulate) {
-			return validate.Usage("The {0} option requires an EtherScan API key.", "--articulate")
+			return validate.Usage("The {0} option requires an Etherscan API key.", "--articulate")
 		}
 
 		if len(opts.Filter) > 0 {