+++ src/apps/chifra/pkg/types/types_transaction.go
@@ -44,24 +44,20 @@ type SimpleTransaction struct {
 	BlockHash            common.Hash     `json:"blockHash"`
 	BlockNumber          Blknum          `json:"blockNumber"`
 	TransactionIndex     uint64          `json:"transactionIndex"`
-	Nonce                uint64          `json:"nonce"`
+	Nonce                uint64          `json:"nonce,omitempty"`
 	Timestamp            Timestamp       `json:"timestamp"`
 	From                 Address         `json:"from"`
 	To                   Address         `json:"to"`
 	Value                Wei             `json:"value"`
-	ExtraValue1          Wei             `json:"extraValue1"`
-	ExtraValue2          Wei             `json:"extraValue2"`
 	Gas                  Gas             `json:"gas"`
 	GasPrice             Gas             `json:"gasPrice"`
 	GasUsed              Gas             `json:"gasUsed"`
 	GasCost              Gas             `json:"gasCost"`
-	MaxFeePerGas         Gas             `json:"maxFeePerGas"`
-	MaxPriorityFeePerGas Gas             `json:"maxPriorityFeePerGas"`
+	MaxFeePerGas         Gas             `json:"maxFeePerGas,omitempty"`
+	MaxPriorityFeePerGas Gas             `json:"maxPriorityFeePerGas,omitempty"`
 	Input                string          `json:"input"`
 	IsError              bool            `json:"isError,omitempty"`
 	HasToken             bool            `json:"hasToken,omitempty"`
-	Cachebits            uint8           `json:"cachebits"`
-	Reserved2            uint8           `json:"reserved2"`
 	Receipt              *SimpleReceipt  `json:"receipt"`
 	Traces               []SimpleTrace   `json:"traces"`
 	ArticulatedTx        *SimpleFunction `json:"articulatedTx,omitempty"`
@@ -143,14 +139,22 @@ func (s *SimpleTransaction) Model(showHidden bool, format string, extraOptions m
 
 	if format == "json" {
 		model["blockHash"] = s.BlockHash
-		model["nonce"] = s.Nonce
+		if s.Nonce > 0 {
+			model["nonce"] = s.Nonce
+		}
 		model["value"] = s.Value.String()
 		model["gas"] = s.Gas
 
 		// TODO: this value could be created when RPC is queried and cached
-		model["ether"] = utils.WeiToEther(&s.Value)
-		model["maxFeePerGas"] = s.MaxPriorityFeePerGas
-		model["maxPriorityFeePerGas"] = s.MaxPriorityFeePerGas
+		// if s.Value > 0 {
+		model["ether"] = utils.WeiToEther(&s.Value).Text('f', 18)
+		// }
+		if s.MaxFeePerGas > 0 {
+			model["maxFeePerGas"] = s.MaxFeePerGas
+		}
+		if s.MaxPriorityFeePerGas > 0 {
+			model["maxPriorityFeePerGas"] = s.MaxPriorityFeePerGas
+		}
 		model["input"] = s.Input
 		if s.HasToken {
 			model["hasToken"] = s.HasToken
@@ -164,7 +168,7 @@ func (s *SimpleTransaction) Model(showHidden bool, format string, extraOptions m
 			model["time"] = date.Format("15:04:05") + " UTC"
 		}
 
-		model["receipt"] = nil
+		// model["receipt"] = nil
 		if s.Receipt != nil {
 			contractAddress := s.Receipt.ContractAddress.Hex()
 
@@ -265,6 +269,9 @@ func (s *SimpleTransaction) Model(showHidden bool, format string, extraOptions m
 
 // EXISTING_CODE
 func (s *SimpleTransaction) SetGasCost(receipt *SimpleReceipt) Gas {
+	if receipt == nil {
+		return 0
+	}
 	s.GasCost = s.GasPrice * receipt.GasUsed
 	return s.GasCost
 }